test: extract prepareProtocolHandler
